Xu_Li_Anh/requirements.txt
Pillow
opencv-python
pytesseract
fpdf